Allows PGBINDIR in check_postgresrc
This is the same approach as it is with PSQL variable (the $NO_PSQL_OPTION
variable har-coded *MUST* be set to 0 to use them).

PGBINDIR can be set in ENV, hard-coded in check_postgres.pl or set via
command-line argument.
PGCONTROLDATA and PSQL can still be used but should be deprecated.

the logic is if NO_PSQL_OPTION=1 :
 * that PGBINDIR can be set via environment variable, but not via config
file or command-line argument.
 * that PSQL can not be set explicitly, but derived from PGBINDIR