Make sure hot_standby_delay doesn't return negative values
Before this patch, the hot_standby_delay action could return negative values.
To minimize the risk, we first query the slave, then the master. And to make
sure, we never get negative values, we set them to zero if we have a negative
value.

Report and ideas from Filip Rembiałkowski.