Make tests running against Postgres 13
I gave a try to the tests and noticed a few failing.  This fixes them so
"make test" runs successfully against Postgres 13.2 using the "postgres" user
under CentOS 8.

A few of the tests were expecting "No matching entries found" when called
with "--includeuser postgres --includeuser mycatbeda" and failing because
"postgres" owned objects were appearing in the result.  I removed
"--includeuser postgres" option so they pass.

And there were 2 problems on the same_schema tests.  Language change message
was not matching with the expected because the check was reporting multiple
differences not just the language.  I changed the pattern to match with this.